I decided to hand this program off to my dearest, challenging him to break it. And BOY HOWDY did he break it.
Like a lot.

He beat this program to death, and in the process found many bugs for me to fix,
One of which was a bug that caused the entire program to crash if the user
tried to use something other than a number as a delete command.
This commit fixes that bug, a bug that caused the error message to display
even when no error was present.

Lastly, I altered visuals to make the program more user-friendly.